breaking updates
1. Remove redundant extern functions (breaking change for GoBattleSim-Python)
2. change several int to unsigned to clear some warnings
3. rename some files to be portable
4. other major fixes

+++ src/Application.cpp
@@ -2,6 +2,7 @@
 #include "Application.h"
 
 #include <stdexcept>
+#include <stdio.h>
 
 namespace GoBattleSim
 {
@@ -41,7 +42,7 @@ void GoBattleSimApp::prepare(const SimInput &input)
     }
     else
     {
-        sprintf(err_msg, "unknown battle mode (%d)", m_sim_input.mode);
+        sprintf(err_msg, "unknown battle mode (%d)", (int)m_sim_input.mode);
         throw std::runtime_error(err_msg);
     }
 }
